Added an ErrorCode to Readable Error Message Translation

+++ frontend/src/app/ui/ticket-tags/ticket-tags.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Output, EventEmitter, OnInit, ViewContainerRef } from '@angular/core';
 import { ApiCallService } from '../../service';
 import { TickettagApi } from '../../api/api/TickettagApi';
 import { TickettaggroupApi } from '../../api/api/TickettaggroupApi';
@@ -8,6 +8,7 @@ import { CreateTicketTagGroupRequestJson } from '../../api/model/CreateTicketTag
 import { UpdateTicketTagGroupRequestJson } from '../../api/model/UpdateTicketTagGroupRequestJson';
 import { TicketTagGroupResultJson } from '../../api/model/TicketTagGroupResultJson';
 import { Observable } from 'rxjs';
+import { Overlay } from 'angular2-modal';
 import { Modal } from 'angular2-modal/plugins/bootstrap';
 
 @Component({
@@ -76,27 +77,27 @@ export class TicketTagsComponent implements OnInit {
     }
 
     onDeleteClicked(tag: TicketTagResultJson) {
-        /*     this.modal.confirm()
-                 .size('sm')
-                 .isBlocking(true)
-                 .showClose(false)
-                 .body('Are you sure you that you want to delete this item?')
-                 .okBtn('Delete')
-                 .open()
-                 .then(a => {
-                     a.result.then(result => {
-                         // Delete clicked
-     */
-        this.apiCallService
-            .call<any>(h => this.ticketTagApi.deleteTicketTagUsingDELETEWithHttpInfo(tag.id, h))
-            .subscribe(param => {
-                this.refresh().subscribe();
-            }
-            );
-        /*      }).catch(result => {
-                  // Cancel clicked
-              });
-          });*/
+        this.modal.confirm()
+            .size('sm')
+            .isBlocking(true)
+            .showClose(false)
+            .body('Are you sure you that you want to delete this item?')
+            .okBtn('Delete')
+            .open()
+            .then(a => {
+                a.result.then(result => {
+                    // Delete clicked
+
+                    this.apiCallService
+                        .call<any>(h => this.ticketTagApi.deleteTicketTagUsingDELETEWithHttpInfo(tag.id, h))
+                        .subscribe(param => {
+                            this.refresh().subscribe();
+                        }
+                        );
+                }).catch(result => {
+                    // Cancel clicked
+                });
+            });
     }
 
     onStartCreate() {
